Updated - using game color menu
- Paint gun now uses the game's color picker menu, and removed '/pg
default' command.
NOTE: Using /pg pick, /pg hsv/rgb or Shift+[LG] with the paint gun will
only hold that color TEMPORARILY, it won't be set in the game's color
picker menu because there's no way for me to do that.
If you want to get colors directly in the color picker menu and the
paint gun, switch over to any cube and use Shift+[LG] with that.
Once the API permits me to set the color to the game's color picker,
then it will make more sense.
- Improved mod's network packets by making them smaller so they use less
bandwidth.
- Added red selection box for invalid block selection.
- Added selection box for symmetry to indicate what blocks are painted.
- Tweaked paint gun animations for first person, 3rd person and
shooting.
- Tweaked spray particles to be more colorful right out the nozzle.
- Added HEX support for '/pg rgb' command, for example: /pg rgb #FF00FF
- Changed the command list when typing '/pg' to a popup window instead
of bundling it in the chat.
- Updated the back-end input handler with the one I use in control
module and ladder to keep things clean and consistent, if you had custom
binds in the config please go and re-make them as they're reset (and
differently syntaxed).
- Fixed green selection box not showing when color picking or when the
color is the same.
- Fixed symmetry painting not working properly in MP because symmetry
planes aren't synchronized by the game.
- Fixed symmetry painting not allowing you to paint if the selected
block is the same color but not the symmetry ones.