Fixed v01.147 compatibility + new features
- Fixed v01.147 compatibility which affected the following things:
-- New issue: input is detected when in menus (except chat);
-- New issue: MP clients will have some problems with colors picked by
the paint gun, to avoid that use the vanilla color picker to replace
colors in your palette.
-- Crosshair no longer changes color depending in what you're aiming at
with the paint gun;

- New way of selecting colors, the paint gun now selects colors
individually from the game's palette (due to the whitelist changes) but
this also means you swap colors way faster.
To swap colors equip the paint gun and use the scroll on the mouse or
next/prev camera keys (default [ and ] ).
- Added a paint palette in the HUD that indicates the selected color and
the other colors as well.
- Added replace-color mode for creative only, shift+R by default. It
replaces the color you click on with the color you selected on the
entire grid or grid system.
- When in color-pick mode, aiming at a player now shows a box around
them to show that you've slected the player and not a block.
- Added various new configurable settings in the config.
- Added a label on the paint gun's ammo bottle model which gets colored
along with the cap when being held.
- Optimized network packets sent by this mod.
- Removed DX9 textures as they wasted space.