Updated
- Changed defaults for palette position and scale.
CONFIG NOTE: if you changed palette HUD position or palette scale in the
config the new defaults will not affect you. My new defaults are
recommended and you can use them by deleting the changed lines in the
config or the entire config, and then /pg reload in chat if you're
already in the game.
- Added a background to fit with the vanilla HUD.
- Added a PaletteBackgroundOpacity setting in the config because the API
is incomplete and I can't read the game's HUD opacity value.
- Fixed PaletteScale config option not working.