Bulk update
- Replaced notifications with right-side UI.
If textAPI is installed it will show a custom drawn UI, otherwise it
shows the vanilla block info UI which can't render the colors.
- Added instant pick color option in the config, defaults to
shift+rightclick.
- Changed how the gun is handled, disabled its vanilla blank-fire and
replaced with a custom firing and sync method.
This allows the following fixes:
--- Removed paintgun screen jittering.
--- Fixed paint gun not usable if weapons are disabled in world.
- Fixed spray sound moving around while painting in first person.
- Changed the network packets to use protobuf.
- General restructure of the code.
- Fixed color conversions and comparisons.
- Fixed paint gun spray particles vanishing when game is paused.
- Space master creative tools mode can now use replace color mode,
instant painting and symmetry without needing to be in a creative
gamemode.